backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.23
pydantic==2.7.3
pydantic-settings==2.1.0
python-multipart==0.0.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
aiofiles==23.2.1
torch==2.3.1
torchvision==0.18.1
facenet-pytorch==2.5.3
Pillow==10.3.0
